Make tests pass on go 1.14 on Windows as well.
Disable a puzzling test for now.

Some semantics of `os` package has changed on Windows:
  https://github.com/golang/go/issues/26295

R=iannucci@chromium.org, nodir@chromium.org
BUG=1056010

+++ common/system/filesystem/filesystem_test.go
@@ -252,9 +252,12 @@ func TestRemoveAll(t *testing.T) {
 
 				Convey(`Can remove the directory when it is read-only`, func() {
 					// Make the directory read-only, and assert that classic os.RemoveAll
-					// fails.
+					// fails. Except on Windows it doesn't, see:
+					// https://github.com/golang/go/issues/26295.
 					So(MakeReadOnly(subDir, nil), ShouldBeNil)
-					So(os.RemoveAll(subDir), ShouldNotBeNil)
+					if runtime.GOOS != "windows" {
+						So(os.RemoveAll(subDir), ShouldNotBeNil)
+					}
 
 					So(RemoveAll(subDir), ShouldBeNil)
 					So(isGone(subDir), ShouldBeTrue)
